refactor: created MinimizeStructure
