+++ pysource_minimize/_utils.py
@@ -1,7 +1,7 @@
 import sys
 
 
-if sys.version_info >= (3, 8):
+if sys.version_info >= (3, 9):
     from ast import unparse
 else:
     from astunparse import unparse  # type: ignore
